Update runtest.cmd
Added variable for the test DLL: Introduced a variable called TEST_DLL to store the path of the test DLL. This allows for easier modification of the DLL path if needed.

Improved quoting: Enclosed the %TEST_DLL% variable in quotes when passing it to vstest.console. This ensures that the path is correctly handled even if it contains spaces or special characters.